Fix StartTimeProvider.setStartTime() on Android (#79)
* Fix StartTimeProvider.setStartTime() on Android

The current way of determining `startTime` is to get `Process.getStartUptimeMillis()`, which is the device uptime at the time when the process was started. 
This is wrong because `startTime` should be a timestamp just like `endTime` which we set to `System.currentTimeMillis()`.

The correct way is to determine `startTime` is to do `SystemClock.uptimeMillis() - Process.getStartUptimeMillis()` to know the elapsed time since the process started, and then subtract that value to `endTime`.

The impact of this bug is limited since there is a check on the next line to use a fallback if the difference between start and end times is greater than one minute, which is always the case. But relying on `SystemClock.uptimeMillis() - Process.getStartUptimeMillis()` is more accurate than the `Process.getElapsedCpuTime()` fallback, because gives us the real elapsed time and not CPU time.

* Improve code

* Simplify setStartTime() by using only Process.getElapsedCpuTime()